+++ docs/pages/components/carousel/examples/ExWithList.vue
@@ -1,0 +1,96 @@
+<template>
+    <b-carousel
+        :autoplay="false"
+        with-carousel-list
+        :indicator="false"
+        :overlay="gallery">
+        <b-carousel-item v-for="(item, i) in items" :key="i">
+            <figure @click="switchGallery(true)" class="image">
+                <img :src="item.image">
+            </figure>
+        </b-carousel-item>
+        <span v-if="gallery" @click="switchGallery(false)" class="modal-close is-large"/>
+        <template slot="list" slot-scope="props">
+            <b-carousel-list
+                v-model="props.active"
+                :data="items"
+                :config="al"
+                :refresh="gallery"
+                @switch="props.switch($event, false)"
+                as-indicator />
+        </template>
+        <template slot="overlay">
+            <div class="has-text-centered has-text-white">
+                Hello i'am overlay!
+            </div>
+        </template>
+    </b-carousel>
+</template>
+
+<script>
+export default {
+    data() {
+        return {
+            gallery: false,
+            al: {
+                hasGrayscale: true,
+                itemsToShow: 2,
+                breakpoints: {
+                    768: {
+                        hasGrayscale: false,
+                        itemsToShow: 4
+                    },
+                    960: {
+                        hasGrayscale: true,
+                        itemsToShow: 6
+                    }
+                }
+            },
+            items: [
+                {
+                    title: 'Slide 1',
+                    image: 'https://picsum.photos/id/0/1230/500'
+                },
+                {
+                    title: 'Slide 2',
+                    image: 'https://picsum.photos/id/1/1230/500'
+                },
+                {
+                    title: 'Slide 3',
+                    image: 'https://picsum.photos/id/2/1230/500'
+                },
+                {
+                    title: 'Slide 4',
+                    image: 'https://picsum.photos/id/3/1230/500'
+                },
+                {
+                    title: 'Slide 5',
+                    image: 'https://picsum.photos/id/4/1230/500'
+                },
+                {
+                    title: 'Slide 6',
+                    image: 'https://picsum.photos/id/5/1230/500'
+                },
+                {
+                    title: 'Slide 7',
+                    image: 'https://picsum.photos/id/6/1230/500'
+                },
+                {
+                    title: 'Slide 8',
+                    image: 'https://picsum.photos/id/7/1230/500'
+                }
+            ]
+        }
+    },
+    methods: {
+        switchGallery(value) {
+            this.gallery = value
+            if (value) {
+                return document.documentElement.classList.add('is-clipped')
+            } else {
+                return document.documentElement.classList.remove('is-clipped')
+            }
+        }
+    }
+}
+</script>