+++ src/components/datetimepicker/Datetimepicker.vue
@@ -22,6 +22,7 @@
         :range="false"
         :disabled="disabled"
         :mobile-native="mobileNative"
+        :focusable="focusable"
         @change-month="$emit('change-month', $event)"
         @change-year="$emit('change-year', $event)">
         <nav class="level is-mobile">
@@ -41,6 +42,7 @@
                     :max-time="maxTime"
                     :size="timepickerSize"
                     :disabled="timepickerDisabled"
+                    :focusable="focusable"
                 />
             </div>
             <div
@@ -130,6 +132,10 @@ export default {
         tzOffset: {
             type: Number,
             default: 0
+        },
+        focusable: {
+            type: Boolean,
+            default: true
         }
     },
     data() {