+++ src/components/tabs/TabItem.vue
@@ -14,7 +14,8 @@ export default {
     data() {
         return {
             isActive: false,
-            transitionName: null
+            transitionName: null,
+            _isTabItem: true // Used internally by Tab
         }
     },
     methods: {
@@ -43,13 +44,10 @@ export default {
             this.$destroy()
             throw new Error('You should wrap bTabItem on a bTabs')
         }
-        this.$parent.tabItems.push(this)
+        this.$parent.refreshSlots()
     },
     beforeDestroy() {
-        const index = this.$parent.tabItems.indexOf(this)
-        if (index >= 0) {
-            this.$parent.tabItems.splice(index, 1)
-        }
+        this.$parent.refreshSlots()
     },
     render(createElement) {
         // if destroy apply v-if