+++ src/components/toast/index.js
@@ -1,9 +1,10 @@
-import Vue from 'vue'
 import Toast from './Toast'
 
-import config from '../../utils/config'
+import config, { VueInstance } from '../../utils/config'
 import { merge } from '../../utils/helpers'
 import { use, registerComponentProgrammatic } from '../../utils/plugins'
+
+let localVueInstance
 
 const ToastProgrammatic = {
     open(params) {
@@ -23,7 +24,7 @@ const ToastProgrammatic = {
         }
         const propsData = merge(defaultParam, params)
 
-        const vm = typeof window !== 'undefined' && window.Vue ? window.Vue : Vue
+        const vm = typeof window !== 'undefined' && window.Vue ? window.Vue : localVueInstance || VueInstance
         const ToastComponent = vm.extend(Toast)
         return new ToastComponent({
             parent,
@@ -35,6 +36,7 @@ const ToastProgrammatic = {
 
 const Plugin = {
     install(Vue) {
+        localVueInstance = Vue
         registerComponentProgrammatic(Vue, 'toast', ToastProgrammatic)
     }
 }