+++ src/utils/TimepickerMixin.js
@@ -130,7 +130,11 @@ export default {
         openOnFocus: Boolean,
         enableSeconds: Boolean,
         defaultMinutes: Number,
-        defaultSeconds: Number
+        defaultSeconds: Number,
+        focusable: {
+            type: Boolean,
+            default: true
+        }
     },
     data() {
         return {
@@ -184,6 +188,7 @@ export default {
         },
 
         minutes() {
+            if (!this.incrementMinutes || this.incrementMinutes < 1) throw new Error('Minute increment cannot be null or less than 1.')
             const minutes = []
             for (let i = 0; i < 60; i += this.incrementMinutes) {
                 minutes.push({
@@ -195,6 +200,7 @@ export default {
         },
 
         seconds() {
+            if (!this.incrementSeconds || this.incrementSeconds < 1) throw new Error('Second increment cannot be null or less than 1.')
             const seconds = []
             for (let i = 0; i < 60; i += this.incrementSeconds) {
                 seconds.push({
@@ -548,6 +554,14 @@ export default {
             // Esc key
             if (this.$refs.dropdown && this.$refs.dropdown.isActive && event.keyCode === 27) {
                 this.toggle(false)
+            }
+        },
+        /**
+         * Emit 'blur' event on dropdown is not active (closed)
+         */
+        onActiveChange(value) {
+            if (!value) {
+                this.onBlur()
             }
         }
     },